👌 IMPROVE: Disable token before delete it from database
resolved #484

+++ src/main/java/com/glinboy/app/service/impl/TokenServiceImpl.java
@@ -91,7 +91,12 @@ public class TokenServiceImpl implements TokenService {
     @Override
     public void delete(Long id) {
         log.debug("Request to delete Token : {}", id);
-        tokenRepository.deleteById(id);
+        Token token = this.tokenRepository.findById(id).orElseThrow();
+        if (Instant.now().isBefore(token.getDeprecateAt())) {
+            this.tokenRepository.updateTokenStatus(id, Boolean.TRUE);
+        } else {
+            tokenRepository.deleteById(id);
+        }
     }
 
     @Override
